refactor(v010): carve voicegw smoke-test out of _legacy
Eighth and largest per-command carve-out in section 2.

Adds voicegateway/cli/smoke_test.py (392 LOC) owning the entire
smoke-test surface: the 'smoke-test' command itself, its three
private helpers (_smoke_active_project, _smoke_pick_models,
_print_smoke_report), the two async runners
(_run_smoke_pipeline_checks for the structural pipeline checks,
_run_smoke_health_checks for the optional --live provider probes),
and the _SMOKE_MODALITIES lookup tuple.

The TODO sub-bullet originally listed only the three named helpers,
but the runners and the _SMOKE_MODALITIES constant are part of the
same surface and reference each other; all five travel together.

Implementation note: the region was ~400 LOC, too large for a clean
Edit old_string. Used a Python one-shot to slice the legacy file by
content boundary (start anchor: '@app.command(name="smoke-test")',
stop anchor: '@app.command(name="serve")'). 14,612 bytes removed.

Drops 'from rich.table import Table' from _legacy.py: smoke-test's
_print_smoke_report was the last consumer left in that file.

13 commands still register on the shared app; pytest tests/test_cli.py
46/46 pass; ruff and mypy clean across the cli package (9 source
files now).
